kernel/fs: Handle root vnode ownership correctly and add more assertions.
In 5b14757a30be0b805e2688cdc5341177f2d703fb (2005), a change was made
to have root nodes be owned by filesystems, rather than
by the VFS (as BeOS did.) However, later on,
in fe5928847a347b3a46eea68628388e60796e71db (2011),
the adding of covered vnodes broke that.

As put_vnode's error was silently ignored, this went unnoticed
until this week, when the new asserts started firing.

So, to fix this, increment the reference count of the root when
setting its covering vnode, and handle the root specially.
Assert in ~fs_mount that there are no vnodes left; if the
filesystem failed to free it, this assert will trip.

Fixes the new assert seen in #19642.