condition_variable: fix NotifyOne possibly not notifying
The notify loop normally checks all pending condition variable entries
and notifies either all of them, or only the first one.

In the case where the first one is being removed from the condition
variable just as the notification happens, it is skipped, however, in
the case of NotifyOne, that should lead to attempting to notify the next
one.

It looks like an oversight in 6d3065508f951d5a0146a842d10a48d83f30def9,
where the notified threads counter was changed in a similar way, but
these two lines of code were not.

The problem seems pretty unlikely, as it would need:
- Multiple threads waiting on the same condition variable,
- The first one being removed just as the variable is notified

Change-Id: I0492c56b15b6b2871954e9114df046ec94539971
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9677
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>