input preflet: Fix tap slider limit values + license update
It was previously using spacing which seems to be a problem.
Spacing value is 11 which by the code will end up on 22 and
the default value is 20 making the tap slided to be very
close to the end with the old code and might be different
in case control looks or other external from the preflet UI
changes in the future.

To avoid that issue the limit has been put of a sensible value
of 50 and divided on 11 marks to visually be divided in chunks
of 10% sized regions.

Change-Id: Ief7461474f0f37eb6fa4bdea1ef8004a531328df
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9901
Reviewed-by: nephele nephele <nep-git@packageloss.eu>