partitioning_systems: Move Intel-related classes back to the "intel" folder.
This partially reverts commit 61790bdb8ab3ab2308f09725767e2059076cd406.

These aren't really "common", they're specific to the "Intel" partitioning
system, which GPT also interacts with. So, move them back to the "intel"
folder, but keep the static library setup the commit moving them
to the "common" folder created, which is indeed better and was most
of the advantage of that commit.