Tracker: Fix minor logic error in Paste item disabled status
... and exclude some more folders from Paste. Gating is
consolidated in "Paste into folder".

Use TargetIsReadOnly() to get parent, not selection.

We want the opposite logic compared to "Paste into folder", we
want to ignore the selected folder or volume's read-only status
and get the read-only status of the parent.

For example File => Paste on config volume in home should be
allowed even though config is a read-only volume. We're _not_
pasting to config, we're pasting to home which is writable.

We want the opposite logic of "Paste in folder" (selection).

Change-Id: Ib13b3a007b86b8d355c994f252f7beb320494c8c
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9895
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>
Reviewed-by: John Scipione <jscipione@gmail.com>
Haiku-Format: Haiku-format Bot <no-reply+haikuformatbot@haiku-os.org>