input mouse mm: Introduce helper funtions for finger handling
It fixes devices that do not store finger flags on fingerWith field
such as Elantech touchpad devices. I think only Synaptic touchpad
use those.

Added are TODO items to change the inner working of these funtions
after a future refactorisation of input devices to handle those flags
and efective finger width values as separate entities.
Commented is "nFingers" with meating of number of fingers which is
expeted to be different from "fingers" which is a finger bitmap
where each bit represent an identified finger.

Change-Id: I5af5d5f86a5ea4f4f06169849afe99395a5e952b
Reviewed-on: https://review.haiku-os.org/c/haiku/+/9913
Reviewed-by: Adrien Destugues <pulkomandy@pulkomandy.tk>