kernel/vm: Add missing Put() when removing a reserved area.
This code is triggered whenever an area gets resized to be as large
as, or larger than, a reservation. That can happen with the new heap
when it grows to rather large sizes, in which case we would then leak
the address space when the team was deleted (including all the page
table memory.)

Fixes memory leaks seen e.g. when running Rust compiles, and especially
other short-lived but high-memory-use applications.